feat: show-redirections-links-in-feature-health-tab
