+++ frontend/web/components/EditHealthProvider.tsx
@@ -14,6 +14,7 @@ import {
 import { components } from 'react-select'
 import InfoMessage from './InfoMessage'
 import InteractiveDemo from './InteractiveDemo'
+import FeatureHealthProviderDocumentationNote from './feature-health/components/FeatureHealthProviderDocumentationNote'
 
 type EditHealthProviderType = {
   projectId: number
@@ -147,32 +148,7 @@ const EditHealthProvider: FC<EditHealthProviderType> = ({
           Learn about Feature Health.
         </Button>
       </p>
-      <InfoMessage>
-        <div>
-          <strong>
-            Follow the documentation to configure alerting using the supported
-            providers.
-          </strong>
-        </div>
-        <div>
-          <span>
-            Sample provider:{' '}
-            <a href='https://docs.flagsmith.com/advanced-use/feature-health#sample-provider'>
-              https://docs.flagsmith.com/advanced-use/feature-health#sample-provider
-            </a>
-          </span>
-        </div>
-        <div>
-          <span>
-            Grafana provider:{' '}
-            <a href='https://docs.flagsmith.com/integrations/apm/grafana/#in-grafana-1'>
-              {' '}
-              https://docs.flagsmith.com/integrations/apm/grafana/#in-grafana-1
-            </a>
-          </span>
-        </div>
-      </InfoMessage>
-
+      <FeatureHealthProviderDocumentationNote />
       <label>Provider Name</label>
       <CreateHealthProviderForm projectId={projectId} />
       <hr className='py-0 my-4' />
