+++ frontend/web/components/modals/CreateFlag.js
@@ -39,7 +39,7 @@ import FeatureHistory from 'components/FeatureHistory'
 import WarningMessage from 'components/WarningMessage'
 import { getPermission } from 'common/services/usePermission'
 import { getChangeRequests } from 'common/services/useChangeRequest'
-import FeatureHealthTabContent from './FeatureHealthTabContent'
+import FeatureHealthTabContent from 'components/feature-health/FeatureHealthTabContent'
 import { IonIcon } from '@ionic/react'
 import { warning } from 'ionicons/icons'
 import FeaturePipelineStatus from 'components/release-pipelines/FeaturePipelineStatus'
