+++ frontend/web/components/users-permissions/OrganisationUsersTable/OrganisationUsersTable.tsx
@@ -28,7 +28,7 @@ const OrganisationUsersTable: React.FC<OrganisationUsersTableProps> = ({
         ? `${user.first_name} ${user.last_name}`
         : `${user.email}`,
       <div>
-        <Tabs uncontrolled hideNavOnSingleTab>
+        <Tabs uncontrolled hideNavOnSingleTab className='mt-4 ml-3'>
           {user.role !== 'ADMIN' && (
             <TabItem tabLabel='Permissions'>
               <div className='pt-4'>
@@ -55,7 +55,7 @@ const OrganisationUsersTable: React.FC<OrganisationUsersTableProps> = ({
     openModal(
       getUserDisplayName(user),
       <div>
-        <Tabs uncontrolled hideNavOnSingleTab>
+        <Tabs uncontrolled hideNavOnSingleTab className='mt-4 ml-3'>
           <TabItem tabLabel='Permissions'>
             <div className='pt-4'>
               <InspectPermissions
